Refactors the rbac_processor to consolidate behavior and simply testing
* removes modules named 'common'
* relocates 'common' features into specific modules
* establishes validators, state accessors, message (proto) accessors
* assigns unique names to functions
* accesses functions through their parent modules to avoid ambiguity
