+++ bin/rbac-tp
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-# Copyright 2017 Intel Corporation
+# Copyright 2018 Contributors to Hyperledger Sawtooth
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
