+++ ledger_sync/rbac_ledger_sync/database.py
@@ -1,4 +1,4 @@
-# Copyright 2017 Intel Corporation
+# Copyright 2018 Contributors to Hyperledger Sawtooth
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -15,7 +15,6 @@
 
 import logging
 import rethinkdb as r
-
 
 LOGGER = logging.getLogger(__name__)
