+++ processor/rbac_processor/main.py
@@ -1,4 +1,4 @@
-# Copyright 2017 Intel Corporation
+# Copyright 2018 Contributors to Hyperledger Sawtooth
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -20,8 +20,7 @@ import os
 from sawtooth_sdk.processor.core import TransactionProcessor
 from sawtooth_sdk.processor.log import init_console_logging
 
-
-from rbac_processor.handler import RBACTransactionHandler
+from rbac_processor.event_handler import RBACTransactionHandler
 
 HOST = os.getenv("HOST", "localhost")
 
@@ -48,14 +47,10 @@ def main(args=None):
     processor = None
     try:
         processor = TransactionProcessor(url=opts.validator_endpoint)
-
         init_console_logging(verbose_level=opts.verbosity)
-
-        handler = RBACTransactionHandler()
-
-        processor.add_handler(handler)
-
+        processor.add_handler(RBACTransactionHandler())
         processor.start()
+
     except KeyboardInterrupt:
         pass
     except Exception as exe:  # pylint: disable=broad-except
