+++ processor/rbac_processor/task/task_validator.py
@@ -1,3 +1,7 @@
+# Copyright 2018 Contributors to Hyperledger Sawtooth
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
 #     http://www.apache.org/licenses/LICENSE-2.0
@@ -9,27 +13,25 @@
 # limitations under the License.
 # -----------------------------------------------------------------------------
 
+from rbac_addressing import addresser
+from rbac_processor import message_accessor
+from rbac_processor import proposal_validator
+from rbac_processor import state_accessor
+from rbac_processor.task import task_validator
+from rbac_processor.user import user_validator
 from sawtooth_sdk.processor.exceptions import InvalidTransaction
 
-from rbac_addressing import addresser
 
-from rbac_processor.common import get_state_entry
-from rbac_processor.common import get_prop_from_container
-from rbac_processor.common import get_task_rel_from_container
-from rbac_processor.common import get_user_from_container
-from rbac_processor.common import is_in_task_rel_container
-from rbac_processor.common import proposal_exists_and_open
-from rbac_processor.common import return_task_rel_container
-from rbac_processor.common import return_prop_container
-from rbac_processor.common import return_user_container
-from rbac_processor.common import validate_identifier_is_user
-from rbac_processor.common import validate_identifier_is_task
+def validate_identifier_is_task(state_entries, identifier, address):
+    try:
+        container = message_accessor.get_task_container(
+            state_accessor.get_state_entry(state_entries, address)
+        )
 
-from rbac_processor.protobuf import proposal_state_pb2
-from rbac_processor.protobuf import task_state_pb2
-
-from rbac_processor.state import get_state
-from rbac_processor.state import set_state
+        if not message_accessor.is_in_task_container(container, identifier):
+            raise InvalidTransaction("{} is not a task.".format(identifier))
+    except KeyError:
+        raise InvalidTransaction("{} is not a task".format(identifier))
 
 
 def validate_task_rel_proposal(header, propose, rel_address, state):
@@ -56,21 +58,29 @@ def validate_task_rel_proposal(header, propose, rel_address, state):
         object_id=task_id, related_id=user_id
     )
 
-    state_entries = get_state(
+    state_entries = state_accessor.get_state(
         state, [user_address, task_address, proposal_address, rel_address]
     )
-    validate_identifier_is_user(state_entries, identifier=user_id, address=user_address)
-    user_entry = get_state_entry(state_entries, user_address)
-    user = get_user_from_container(return_user_container(user_entry), user_id)
+    user_validator.validate_identifier_is_user(
+        state_entries, identifier=user_id, address=user_address
+    )
+    user_entry = state_accessor.get_state_entry(state_entries, user_address)
+    user = message_accessor.get_user_from_container(
+        message_accessor.get_user_container(user_entry), user_id
+    )
 
-    validate_identifier_is_task(state_entries, identifier=task_id, address=task_address)
+    task_validator.validate_identifier_is_task(
+        state_entries, identifier=task_id, address=task_address
+    )
 
     try:
-        task_rel_entry = get_state_entry(state_entries, rel_address)
-        task_rel_container = return_task_rel_container(task_rel_entry)
+        task_rel_entry = state_accessor.get_state_entry(state_entries, rel_address)
+        task_rel_container = message_accessor.get_task_rel_container(task_rel_entry)
 
         if (header.signer_public_key not in [user.user_id, user.manager_id]) and (
-            not is_in_task_rel_container(task_rel_container, task_id, user_id)
+            not message_accessor.is_in_task_rel_container(
+                task_rel_container, task_id, user_id
+            )
         ):
             raise InvalidTransaction(
                 "Txn signer {} is not the user or the user's "
@@ -78,7 +88,9 @@ def validate_task_rel_proposal(header, propose, rel_address, state):
                     header.signer_public_key, [user.user_id, user.manager_id]
                 )
             )
-        if is_in_task_rel_container(task_rel_container, task_id, user_id):
+        if message_accessor.is_in_task_rel_container(
+            task_rel_container, task_id, user_id
+        ):
             raise InvalidTransaction(
                 "User {} is already in the Task {} "
                 "relationship".format(user_id, task_id)
@@ -113,28 +125,30 @@ def validate_task_rel_del_proposal(header, propose, rel_address, state):
         object_id=propose.task_id, related_id=propose.user_id
     )
 
-    state_entries = get_state(
+    state_entries = state_accessor.get_state(
         state, [user_address, task_address, proposal_address, rel_address]
     )
 
-    validate_identifier_is_user(
+    user_validator.validate_identifier_is_user(
         state_entries, identifier=propose.user_id, address=user_address
     )
 
-    user_entry = get_state_entry(state_entries, user_address)
+    user_entry = state_accessor.get_state_entry(state_entries, user_address)
 
-    user = get_user_from_container(return_user_container(user_entry), propose.user_id)
+    user = message_accessor.get_user_from_container(
+        message_accessor.get_user_container(user_entry), propose.user_id
+    )
 
-    validate_identifier_is_task(
+    task_validator.validate_identifier_is_task(
         state_entries, identifier=propose.task_id, address=task_address
     )
 
     try:
-        task_rel_entry = get_state_entry(state_entries, rel_address)
-        task_rel_container = return_task_rel_container(task_rel_entry)
+        task_rel_entry = state_accessor.get_state_entry(state_entries, rel_address)
+        task_rel_container = message_accessor.get_task_rel_container(task_rel_entry)
 
         if (header.signer_public_key not in [user.user_id, user.manager_id]) and (
-            not is_in_task_rel_container(
+            not message_accessor.is_in_task_rel_container(
                 task_rel_container, propose.task_id, propose.user_id
             )
         ):
@@ -145,7 +159,7 @@ def validate_task_rel_del_proposal(header, propose, rel_address, state):
                 )
             )
 
-        if not is_in_task_rel_container(
+        if not message_accessor.is_in_task_rel_container(
             task_rel_container, propose.task_id, propose.user_id
         ):
             raise InvalidTransaction(
@@ -189,13 +203,15 @@ def validate_task_admin_or_owner(
     )
 
     if not is_remove:
-        state_entries = get_state(state, [txn_signer_rel_address, proposal_address])
+        state_entries = state_accessor.get_state(
+            state, [txn_signer_rel_address, proposal_address]
+        )
     else:
-        state_entries = get_state(
+        state_entries = state_accessor.get_state(
             state, [txn_signer_rel_address, task_rel_address, proposal_address]
         )
 
-    if not proposal_exists_and_open(
+    if not proposal_validator.proposal_exists_and_open(
         state_entries, proposal_address, confirm.proposal_id
     ):
         raise InvalidTransaction(
@@ -203,14 +219,14 @@ def validate_task_admin_or_owner(
             "is not open".format(confirm.proposal_id)
         )
     try:
-        entry = get_state_entry(state_entries, txn_signer_rel_address)
-        task_rel_container = return_task_rel_container(entry)
+        entry = state_accessor.get_state_entry(state_entries, txn_signer_rel_address)
+        task_rel_container = message_accessor.get_task_rel_container(entry)
     except KeyError:
         raise InvalidTransaction(
             "Signer {} does not have the Task permissions "
             "to close the proposal".format(header.signer_public_key)
         )
-    if not is_in_task_rel_container(
+    if not message_accessor.is_in_task_rel_container(
         task_rel_container, task_id=confirm.task_id, identifier=header.signer_public_key
     ):
         raise InvalidTransaction(
@@ -222,107 +238,20 @@ def validate_task_admin_or_owner(
     return state_entries
 
 
-def handle_propose_state_set(
-    state_entries,
-    header,
-    payload,
-    address,
-    proposal_type,
-    state,
-    related_type="user_id",
-):
-    try:
-
-        entry = get_state_entry(state_entries, address=address)
-        proposal_container = return_prop_container(entry)
-    except KeyError:
-        proposal_container = proposal_state_pb2.ProposalsContainer()
-
-    proposal = proposal_container.proposals.add()
-
-    proposal.proposal_id = payload.proposal_id
-    proposal.object_id = payload.task_id
-    proposal.target_id = getattr(payload, related_type)
-    proposal.proposal_type = proposal_type
-    proposal.status = proposal_state_pb2.Proposal.OPEN
-    proposal.opener = header.signer_public_key
-    proposal.open_reason = payload.reason
-    proposal.metadata = payload.metadata
-
-    set_state(state, {address: proposal_container.SerializeToString()})
-
-
-def handle_confirm(state_entries, header, confirm, task_rel_address, state, is_remove):
-    """ Updates proposal and task relationship objects according to the
-        task admin/owner transaction.
-
-        Args:
-            state_entries: List of states for the proposal, task relationship,
-            and task admins object.
-            header (TransactionHeader): The protobuf TransactionHeader.
-            confirm (RBACPayload): The protobuf RBACPayload.
-            task_rel_address (str): The task relationship address.
-            state (Context): The class that handles state gets and sets.
-            is_remove (boolean): Determines if task admin/owner is being removed or added.
-
-    """
-    proposal_address = addresser.make_proposal_address(
-        object_id=confirm.task_id, related_id=confirm.user_id
-    )
-
-    proposal_entry = get_state_entry(state_entries, proposal_address)
-    proposal_container = return_prop_container(proposal_entry)
-    proposal = get_prop_from_container(
-        proposal_container, proposal_id=confirm.proposal_id
-    )
-
-    proposal.status = proposal_state_pb2.Proposal.CONFIRMED
-    proposal.closer = header.signer_public_key
-    proposal.close_reason = confirm.reason
-
-    address_values = {proposal_address: proposal_container.SerializeToString()}
+def validate_create_task_state(state_entries, payload):
+    user_validator.validate_list_of_user_are_users(state_entries, payload.admins)
+    user_validator.validate_list_of_user_are_users(state_entries, payload.owners)
 
     try:
-        task_rel_entry = get_state_entry(state_entries, task_rel_address)
-        task_rel_container = return_task_rel_container(task_rel_entry)
-    except KeyError:
-        task_rel_container = task_state_pb2.TaskRelationshipContainer()
-
-    try:
-        task_rel = get_task_rel_from_container(
-            container=task_rel_container,
-            task_id=confirm.task_id,
-            identifier=confirm.user_id,
+        entry = state_accessor.get_state_entry(
+            state_entries, addresser.make_task_attributes_address(payload.task_id)
         )
+        container = message_accessor.get_task_container(entry)
+
+        if message_accessor.is_in_task_container(container, payload.task_id):
+            raise InvalidTransaction(
+                "Task with id {} already in " "state".format(payload.task_id)
+            )
     except KeyError:
-        task_rel = task_rel_container.relationships.add()
-        task_rel.task_id = confirm.task_id
-
-    if not is_remove:
-        task_rel.identifiers.append(confirm.user_id)
-    else:
-        task_rel.identifiers.remove(confirm.user_id)
-
-    address_values[task_rel_address] = task_rel_container.SerializeToString()
-
-    set_state(state, address_values)
-
-
-def handle_reject(state_entries, header, reject, state):
-    proposal_address = addresser.make_proposal_address(
-        object_id=reject.task_id, related_id=reject.user_id
-    )
-
-    proposal_entry = get_state_entry(state_entries, proposal_address)
-    proposal_container = return_prop_container(proposal_entry)
-    proposal = get_prop_from_container(
-        proposal_container, proposal_id=reject.proposal_id
-    )
-
-    proposal.status = proposal_state_pb2.Proposal.REJECTED
-    proposal.closer = header.signer_public_key
-    proposal.close_reason = reject.reason
-
-    address_values = {proposal_address: proposal_container.SerializeToString()}
-
-    set_state(state, address_values)
+        # The task container is not in state, so no at this address.
+        pass
