+++ processor/rbac_processor/task/tasks.py
@@ -1,4 +1,4 @@
-# Copyright 2017 Intel Corporation
+# Copyright 2018 Contributors to Hyperledger Sawtooth
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -16,17 +16,14 @@
 from sawtooth_sdk.processor.exceptions import InvalidTransaction
 
 from rbac_addressing import addresser
-from rbac_processor.common import get_state_entry
-from rbac_processor.common import is_in_task_container
-from rbac_processor.common import return_task_container
-from rbac_processor.common import validate_list_of_user_are_users
+from rbac_processor.task import task_validator
 from rbac_processor.protobuf import task_state_pb2
 from rbac_processor.protobuf import task_transaction_pb2
-from rbac_processor.state import get_state
-from rbac_processor.state import set_state
+from rbac_processor import message_accessor
+from rbac_processor import state_accessor
 
 
-def apply_create_task(header, payload, state):
+def new_task(payload, state):
     create_payload = task_transaction_pb2.CreateTask()
     create_payload.ParseFromString(payload.content)
 
@@ -62,45 +59,21 @@ def apply_create_task(header, payload, state):
             ]
         )
 
-    state_entries = get_state(state, addresses)
+    state_entries = state_accessor.get_state(state, addresses)
 
-    validate_create_task_state(state_entries=state_entries, payload=create_payload)
+    task_validator.validate_create_task_state(
+        state_entries=state_entries, payload=create_payload
+    )
 
-    handle_create_task(state_entries, create_payload, state)
-
-
-def validate_create_task_data(payload):
-    if len(payload.name) < 4:
-        raise InvalidTransaction("The task name must be longer " "than 4 characters")
+    create_task(state_entries, create_payload, state)
 
 
-def validate_create_task_state(state_entries, payload):
-
-    validate_list_of_user_are_users(state_entries, payload.admins)
-
-    validate_list_of_user_are_users(state_entries, payload.owners)
-
+def create_task(state_entries, payload, state):
     try:
-        entry = get_state_entry(
+        entry = state_accessor.get_state_entry(
             state_entries, addresser.make_task_attributes_address(payload.task_id)
         )
-        container = return_task_container(entry)
-
-        if is_in_task_container(container, payload.task_id):
-            raise InvalidTransaction(
-                "Task with id {} already in " "state".format(payload.task_id)
-            )
-    except KeyError:
-        # The task container is not in state, so no at this address.
-        pass
-
-
-def handle_create_task(state_entries, payload, state):
-    try:
-        entry = get_state_entry(
-            state_entries, addresser.make_task_attributes_address(payload.task_id)
-        )
-        container = return_task_container(entry)
+        container = message_accessor.get_task_container(entry)
 
     except KeyError:
         container = task_state_pb2.TaskAttributesContainer()
@@ -126,9 +99,8 @@ def handle_create_task(state_entries, payload, state):
     address_values.update(
         _handle_task_rel_container(
             state_entries=state_entries,
-            create_task=payload,
+            task_id=payload.task_id,
             pubkeys_by_address=pubkeys_by_address,
-            state=state,
         )
     )
 
@@ -145,9 +117,8 @@ def handle_create_task(state_entries, payload, state):
     address_values.update(
         _handle_task_rel_container(
             state_entries=state_entries,
-            create_task=payload,
+            task_id=payload.task_id,
             pubkeys_by_address=pubkeys_by_address,
-            state=state,
         )
     )
 
@@ -155,21 +126,21 @@ def handle_create_task(state_entries, payload, state):
         addresser.make_task_attributes_address(payload.task_id)
     ] = container.SerializeToString()
 
-    set_state(state, address_values)
+    state_accessor.set_state(state, address_values)
 
 
-def _handle_task_rel_container(state_entries, create_task, pubkeys_by_address, state):
+def _handle_task_rel_container(state_entries, task_id, pubkeys_by_address):
     entries_to_set = {}
 
     for addr, pubkeys in pubkeys_by_address.items():
         try:
-            state_entry = get_state_entry(state_entries, addr)
+            state_entry = state_accessor.get_state_entry(state_entries, addr)
             container = task_state_pb2.TaskRelationshipContainer()
             container.ParseFromString(state_entry.data)
         except KeyError:
             container = task_state_pb2.TaskRelationshipContainer()
 
-        _add_task_rel_to_container(container, create_task.task_id, pubkeys)
+        _add_task_rel_to_container(container, task_id, pubkeys)
 
         entries_to_set[addr] = container.SerializeToString()
